# Conflicts:
#	drivers/flash/spi_nor.c
#	drivers/modem/modem_cellular.c
#	include/zephyr/kernel/mm/demand_paging.h
#	include/zephyr/kernel/thread.h
#	include/zephyr/kernel/thread_stack.h
#	include/zephyr/net/wifi_mgmt.h
#	include/zephyr/pm/device.h
#	kernel/fatal.c
#	kernel/init.c
#	kernel/msg_q.c
#	kernel/poll.c
#	kernel/sem.c
#	kernel/thread.c
#	kernel/timeout.c
#	kernel/usage.c
#	subsys/bluetooth/host/hci_raw.c
#	subsys/modem/modem_cmux.c
#	subsys/net/ip/ipv4.c
#	subsys/net/ip/ipv6.c
#	subsys/net/ip/net_if.c
#	subsys/net/l2/wifi/wifi_shell.c
#	subsys/shell/backends/shell_uart.c

+++ subsys/net/buf.c
@@ -331,7 +331,7 @@ success:
 	buf->ref   = 1U;
 	buf->flags = 0U;
 	buf->frags = NULL;
-	buf->size  = size;
+	buf->size  = (uint16_t)size;
 	net_buf_reset(buf);
 
 #if defined(CONFIG_NET_BUF_POOL_USAGE)
@@ -534,7 +534,7 @@ struct net_buf *net_buf_clone(struct net_buf *buf, k_timeout_t timeout)
 			return NULL;
 		}
 
-		clone->size = size;
+		clone->size = (uint16_t)size;
 		clone->data = clone->__buf + net_buf_headroom(buf);
 		net_buf_add_mem(clone, buf->data, buf->len);
 	}
@@ -677,7 +677,7 @@ size_t net_buf_append_bytes(struct net_buf *buf, size_t len,
 	size_t max_size;
 
 	do {
-		uint16_t count = MIN(len, net_buf_tailroom(frag));
+		size_t count = MIN(len, net_buf_tailroom(frag));
 
 		net_buf_add_mem(frag, value8, count);
 		len -= count;
